Split additional sign tests to real and plan modules
Adding write support for nested content instances to additional sign
api endpoint requires more thorough testing that will bloat the test
files so much that it makes sense to split them to separate plan
tests and real tests.

This is done in separate commit so the diff for new and modified tests
is nicer. This commit itself does not contain any changes to the tests.

Refs LIIK-183
